pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grpo-trainer"
version = "2.0.0"
description = "Advanced GRPO Training Framework for LLM Fine-tuning on Math & Reasoning Datasets"
readme = "README.md"
license = {text = "GPL-3.0"}
authors = [
    {name = "kossisoroyce"}
]
keywords = ["grpo", "reinforcement-learning", "llm", "fine-tuning", "transformers", "math-reasoning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "trl>=0.8.0",
    "peft>=0.10.0",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "bitsandbytes>=0.43.0",
    "wandb>=0.16.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "sentencepiece>=0.1.99",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
flash-attn = [
    "flash-attn>=2.5.0",
]
deepspeed = [
    "deepspeed>=0.14.0",
]
all = [
    "grpo-trainer[dev,flash-attn,deepspeed]",
]

[project.scripts]
grpo-train = "grpo_trainer.cli:train"
grpo-eval = "grpo_trainer.cli:evaluate"

[project.urls]
Homepage = "https://github.com/kossisoroyce/grpo-trainer"
Repository = "https://github.com/kossisoroyce/grpo-trainer"
Issues = "https://github.com/kossisoroyce/grpo-trainer/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=grpo_trainer --cov-report=term-missing"
